requirements.txt
matplotlib==3.5.1
moviepy==1.0.3
Pillow==9.0.0
pytesseract==0.3.8
textdistance==4.2.2
